requirements.txt
Flask==2.3.3
python-docx==0.8.11
Werkzeug==2.3.7 